third_party/blink/web_tests/fast/forms/calendar-picker/calendar-picker-mouse-operations-expected.txt

Tests if calendar picker mouse opearations work as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Check that page popup doesn't exist at first.
PASS document.getElementById("mock-page-popup") is null
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is "2000-01-02"
PASS highlightedValue() is "2000-01-02"
PASS highlightedDayCells() is "2000-01-02"
Check that hovering over an entry highlights it.
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is "2000-01-02"
PASS highlightedValue() is "1999-12-31"
PASS highlightedDayCells() is "1999-12-31"
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is "2000-01-02"
PASS highlightedValue() is "2000-01-04"
PASS highlightedDayCells() is "2000-01-04"
Check that navigating between months maintains highlighted position.
PASS currentMonth() is "2000-02"
PASS selectedValue() is "2000-01-02"
PASS selectedDayCells() is ""
PASS highlightedValue() is "2000-02-08"
PASS highlightedDayCells() is "2000-02-08"
Check that mouse click sets the value.
PASS document.getElementById("date").value is "2000-02-05"
Check that clicking the month popup button opens the month popup.
PASS popupWindow.global.picker.monthPopupView.isVisible is false
PASS popupWindow.global.picker.monthPopupView.isVisible is true
PASS highlightedMonthButton() is "2000-02"
Check that hovering over a month button highlights it.
PASS highlightedMonthButton() is "2000-10"
Check that month wheel scrolls the list in the month popup.
PASS checkYearListViewScrollOffset() > 0 is true
PASS checkYearListViewScrollOffset() < 0 is true
Check that clicking an year list cell opens it.
PASS checkYearListViewScrollOffset() > 0 is true
PASS highlightedMonthButton() is "1999-01"
Check that clicking the month popup sets the month.
PASS popupWindow.global.picker.monthPopupView.isVisible is false
PASS currentMonth() is "1999-06"
PASS successfullyParsed is true

TEST COMPLETE
